backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
azure-ai-formrecognizer==3.2.1
azure-common==1.1.28
azure-core==1.27.1
azure-search==1.0.0b2
azure-storage-blob==12.16.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
dataclasses-json==0.5.9
fake-useragent==1.1.3
Flask==2.3.2
frozenlist==1.3.3
greenlet==2.0.2
hiredis==2.2.3
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.225
langchainplus-sdk==0.0.20
lupa==2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
pycparser==2.21
pydantic==1.10.11
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==3.5.3
redisearch==2.1.1
rediserver==0.1
regex==2023.6.3
rejson==0.5.6
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.6
yarl==1.9.2
python-docx==0.8.11
nltk
--extra-index-url https://pkgs.dev.azure.com/azure-sdk/public/_packaging/azure-sdk-for-python/pypi/simple/
azure-search-documents==11.4.0a20230509004
PyPDF4
PyPDF2
reportlab==4.0.8
pdf2docx
